requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
joblib>=1.3
altair>=5.0
PyYAML>=6.0
